+++ service/redshift/api.go
@@ -37478,6 +37478,9 @@ type Snapshot struct {
 	// The list of node types that this cluster snapshot is able to restore into.
 	RestorableNodeTypes []*string `locationNameList:"NodeType" type:"list"`
 
+	// The Amazon Resource Name (ARN) of the snapshot.
+	SnapshotArn *string `type:"string"`
+
 	// The time (in UTC format) when Amazon Redshift began the snapshot. A snapshot
 	// contains a copy of the cluster data as of this exact time.
 	SnapshotCreateTime *time.Time `type:"timestamp"`
@@ -37695,6 +37698,12 @@ func (s *Snapshot) SetPort(v int64) *Snapshot {
 // SetRestorableNodeTypes sets the RestorableNodeTypes field's value.
 func (s *Snapshot) SetRestorableNodeTypes(v []*string) *Snapshot {
 	s.RestorableNodeTypes = v
+	return s
+}
+
+// SetSnapshotArn sets the SnapshotArn field's value.
+func (s *Snapshot) SetSnapshotArn(v string) *Snapshot {
+	s.SnapshotArn = &v
 	return s
 }